chore(embeddings): refactor
